+++ protected/humhub/modules/content/controllers/MoveController.php
@@ -10,7 +10,7 @@
 namespace humhub\modules\content\controllers;
 
 use HttpException;
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\content\components\ContentContainerController;
 use humhub\modules\content\models\Content;
 use humhub\modules\content\models\forms\MoveContentForm;