+++ protected/humhub/modules/installer/views/config/sample-data.php
@@ -1,8 +1,8 @@
 <?php
 
+use humhub\helpers\Html;
 use humhub\modules\installer\forms\SampleDataForm;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use yii\bootstrap\Html;
+use humhub\widgets\form\ActiveForm;
 
 /* @var SampleDataForm $model */
 ?>
@@ -27,5 +27,3 @@ use yii\bootstrap\Html;
         <?php ActiveForm::end(); ?>
     </div>
 </div>
-
-