+++ protected/humhub/modules/space/widgets/views/spaceNameColorInput.php
@@ -1,9 +1,8 @@
 <?php
 
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\ColorPickerField;
+use humhub\components\View;
+use humhub\widgets\form\ActiveForm;
 use yii\base\Model;
-use yii\widgets\ActiveForm;
 
 /**
  * @var $this View
@@ -19,20 +18,24 @@ if ($model->color === null) {
 }
 ?>
 
-<div id="<?= $containerId ?>" class="form-group space-color-chooser-edit" style="margin-top: 5px;">
-    <?= ColorPickerField::widget(['model' => $model, 'field' => 'color', 'container' => $containerId]); ?>
+<label class="form-label" for="<?= $containerId ?>"><?= $model->attributeLabels()['name'] ?></label>
 
-    <?= $form->field($model, 'name', ['template' => '
-            {label}
-            <div class="input-group">
-                <span class="input-group-addon">
-                    <i></i>
-                </span>
-                {input}
-            </div>
-            {error}{hint}',
-    ])->textInput(array_merge([
-        'placeholder' => Yii::t('SpaceModule.manage', 'Space name'),
-        'maxlength' => 45,
-    ], ($focus ? ['autofocus' => ''] : []))); ?>
+<div id="<?= $containerId ?>" class="mb-3 input-group space-color-chooser-edit">
+
+    <div class="input-group-prepend">
+        <?= $form->field($model, 'color', ['options' => []])
+            ->colorInput(['style' => 'border-top-right-radius: 0; border-bottom-right-radius: 0; height: 36px;'])
+            ->label(false) ?>
+    </div>
+
+    <?= $form->field($model, 'name', ['options' => ['class' => 'flex-grow-1']])
+        ->textInput(array_merge(
+            [
+                'placeholder' => Yii::t('SpaceModule.manage', 'Space name'),
+                'maxlength' => 45,
+                'style' => 'margin-left: calc(-1 * var(--bs-border-width)); border-top-left-radius: 0; border-bottom-left-radius: 0; height: 36px;',
+            ],
+            ($focus ? ['autofocus' => ''] : []),
+        ))
+        ->label(false) ?>
 </div>