+++ protected/humhub/modules/ui/tests/codeception/unit/TabbedFormTest.php
@@ -8,8 +8,8 @@
 
 namespace tests\codeception\unit;
 
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\modules\ui\form\widgets\FormTabs;
+use humhub\widgets\bootstrap\FormTabs;
+use humhub\widgets\form\ActiveForm;
 use tests\codeception\_support\HumHubDbTestCase;
 use tests\codeception\models\TestTabbedFormModel;