+++ protected/humhub/modules/user/models/fieldtype/UserEmail.php
@@ -8,7 +8,7 @@
 
 namespace humhub\modules\user\models\fieldtype;
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\user\models\User;
 
 /**