+++ protected/humhub/modules/user/views/account/delete_spaceowner.php
@@ -1,6 +1,6 @@
 <?php
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\space\widgets\Image;
 
 ?>
@@ -10,12 +10,12 @@ use humhub\modules\space\widgets\Image;
 <?= Yii::t('UserModule.account', 'You are currently the owner of following spaces:'); ?><br/>
 
 <?php foreach ($ownSpaces as $space): ?>
-    <div class="media">
-        <div class="media-left" style="padding-right:6px">
+    <div class="d-flex">
+        <div class="flex-shrink-0 me-2">
             <?= Image::widget(['space' => $space, 'width' => 38, 'link' => true]); ?>
         </div>
-        <div class="media-body">
-            <h4 class="media-heading"><?= Html::containerLink($space); ?></h4>
+        <div class="flex-grow-1">
+            <h4 class="mt-0"><?= Html::containerLink($space); ?></h4>
             <?= Yii::t('SpaceModule.base', '{count} members', ['count' => $space->getMemberships()->count()]); ?>
         </div>
     </div>